Add confirmed guns directory
Guns in this folder give the smallest implementaion of tentatively
suggested guns.

For example the p250 gun that was suggested as record breaker is not
actually a record breaker when the necessary tweaks to it are made.

The 4 guns in the confirmed directory all disappear from the tentative
list and 3 of them go into the main list... but not the p250.